refactor: Remove async `indexer_rules_engine` (#387)
We have two versions of `indexer_rules_engine`; sync and async. This
logic does not need to be `async` so I've removed this version and
refactored so only the sync version is used.

Additionally, I've migrated the `async` tests across to use the `sync`
version, and have added `cargo test` to Github Actions so that they are
always run. This required fixing the `set_provisioning_flag` test, as
well as ignoring the integration tests which rely on AWS access keys.
